extract more utility functions
